+++ database/migrations/2025_09_07_131034_siswa_penghargaan.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('siswapenghargaan', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('nis');
+            $table->bigInteger('nis');
             $table->string('id_penghargaan');
             $table->timestamps();
             $table->foreign('nis')->references('nis')->on('siswa')->onDelete('cascade');
