+++ database/migrations/2025_09_07_131045_siswa_sp.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('siswaperingatan', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('nis');
+            $table->bigInteger('nis');
             $table->string('id_sp');
             $table->timestamps();
              $table->foreign('nis')->references('nis')->on('siswa')->onDelete('cascade');
